+++ src/Domain/Enums/SequenceType.cs
@@ -1,0 +1,17 @@
+namespace Lucy.Domain.Enums;
+
+/// <summary>
+/// The types of sequences.
+/// </summary>
+public enum SequenceType
+{
+    /// <summary>
+    /// A ticket sequence.
+    /// </summary>
+    Ticket = 0,
+
+    /// <summary>
+    /// An iteration sequence.
+    /// </summary>
+    Iteration = 1
+}